fixed a few warnings

+++ src/Game.cpp
@@ -58,7 +58,7 @@ void Game::start()
 			{
 				const float TargetDelta = 0.0167f * index;
 				if (mDeltaTime < TargetDelta)
-					SDL_Delay(static_cast<float>(TargetDelta - mDeltaTime) * 1000.f);
+					SDL_Delay(TargetDelta - mDeltaTime * 1000);
 			}
             mWindow->update();
             doAction(mIManager->processEvents(mWindow->getEvent()));
